Move handlers and timeouts to package
