+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/builder/EphemeralActionableComponentImpl.kt
@@ -4,7 +4,7 @@ import io.github.freya022.botcommands.api.ReceiverConsumer
 import io.github.freya022.botcommands.api.components.builder.EphemeralHandlerBuilder
 import io.github.freya022.botcommands.api.components.builder.IEphemeralActionableComponent
 import io.github.freya022.botcommands.api.core.BContext
-import io.github.freya022.botcommands.internal.components.EphemeralHandler
+import io.github.freya022.botcommands.internal.components.handler.EphemeralHandler
 import net.dv8tion.jda.api.events.interaction.component.GenericComponentInteractionCreateEvent
 
 internal class EphemeralActionableComponentImpl<T : IEphemeralActionableComponent<T, E>, E : GenericComponentInteractionCreateEvent> internal constructor(
