+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/controller/ComponentsListener.kt
@@ -20,15 +20,15 @@ import io.github.freya022.botcommands.api.core.service.getInterfacedServices
 import io.github.freya022.botcommands.api.core.service.getServiceOrNull
 import io.github.freya022.botcommands.api.core.utils.simpleNestedName
 import io.github.freya022.botcommands.internal.commands.ratelimit.withRateLimit
-import io.github.freya022.botcommands.internal.components.ComponentDescriptor
-import io.github.freya022.botcommands.internal.components.ComponentHandlerOption
 import io.github.freya022.botcommands.internal.components.ComponentType
-import io.github.freya022.botcommands.internal.components.EphemeralHandler
 import io.github.freya022.botcommands.internal.components.data.AbstractComponentData
 import io.github.freya022.botcommands.internal.components.data.ComponentGroupData
 import io.github.freya022.botcommands.internal.components.data.EphemeralComponentData
 import io.github.freya022.botcommands.internal.components.data.PersistentComponentData
-import io.github.freya022.botcommands.internal.components.repositories.ComponentHandlerContainer
+import io.github.freya022.botcommands.internal.components.handler.ComponentDescriptor
+import io.github.freya022.botcommands.internal.components.handler.ComponentHandlerContainer
+import io.github.freya022.botcommands.internal.components.handler.ComponentHandlerOption
+import io.github.freya022.botcommands.internal.components.handler.EphemeralHandler
 import io.github.freya022.botcommands.internal.components.repositories.ComponentRepository
 import io.github.freya022.botcommands.internal.core.BContextImpl
 import io.github.freya022.botcommands.internal.core.ExceptionHandler
