+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/data/AbstractComponentData.kt
@@ -3,9 +3,9 @@ package io.github.freya022.botcommands.internal.components.data
 import io.github.freya022.botcommands.api.components.ComponentInteractionFilter
 import io.github.freya022.botcommands.api.components.data.ComponentTimeout
 import io.github.freya022.botcommands.api.components.data.InteractionConstraints
-import io.github.freya022.botcommands.internal.components.ComponentHandler
 import io.github.freya022.botcommands.internal.components.ComponentType
 import io.github.freya022.botcommands.internal.components.LifetimeType
+import io.github.freya022.botcommands.internal.components.handler.ComponentHandler
 
 internal sealed class AbstractComponentData(
     componentId: Int,
