+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/handler/ComponentDescriptor.kt
@@ -1,4 +1,4 @@
-package io.github.freya022.botcommands.internal.components
+package io.github.freya022.botcommands.internal.components.handler
 
 import io.github.freya022.botcommands.api.commands.builder.CustomOptionBuilder
 import io.github.freya022.botcommands.internal.IExecutableInteractionInfo
