+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/handler/ComponentHandlerOption.kt
@@ -1,4 +1,4 @@
-package io.github.freya022.botcommands.internal.components
+package io.github.freya022.botcommands.internal.components.handler
 
 import io.github.freya022.botcommands.api.parameters.resolvers.ComponentParameterResolver
 import io.github.freya022.botcommands.internal.core.options.OptionImpl
