+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/handler/ComponentHandlerOptionBuilder.kt
@@ -1,4 +1,4 @@
-package io.github.freya022.botcommands.internal.components
+package io.github.freya022.botcommands.internal.components.handler
 
 import io.github.freya022.botcommands.api.core.options.builder.OptionBuilder
 import io.github.freya022.botcommands.internal.parameters.OptionParameter
