+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/timeout/EphemeralTimeoutHandlers.kt
@@ -1,8 +1,9 @@
-package io.github.freya022.botcommands.internal.components.repositories
+package io.github.freya022.botcommands.internal.components.timeout
 
 import io.github.freya022.botcommands.api.components.Components
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.api.core.service.annotations.Dependencies
+import io.github.freya022.botcommands.internal.components.EphemeralHandlers
 
 @BService
 @Dependencies(Components::class)
