+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/timeout/GroupTimeoutHandlers.kt
@@ -1,8 +1,8 @@
-package io.github.freya022.botcommands.internal.components.repositories
+package io.github.freya022.botcommands.internal.components.timeout
 
 import io.github.freya022.botcommands.api.components.Components
-import io.github.freya022.botcommands.api.components.annotations.ComponentTimeoutHandler
-import io.github.freya022.botcommands.api.components.data.ComponentTimeoutData
+import io.github.freya022.botcommands.api.components.annotations.GroupTimeoutHandler
+import io.github.freya022.botcommands.api.components.data.GroupTimeoutData
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.api.core.service.annotations.Dependencies
 import io.github.freya022.botcommands.internal.core.reflection.MemberFunction
@@ -14,13 +14,13 @@ import kotlin.reflect.full.findAnnotation
 
 @BService
 @Dependencies(Components::class)
-internal class ComponentTimeoutHandlers(functionAnnotationsMap: FunctionAnnotationsMap) : HandlerContainer {
+internal class GroupTimeoutHandlers(functionAnnotationsMap: FunctionAnnotationsMap) : HandlerContainer {
     private val map: Map<String, MemberFunction<*>> =
-        functionAnnotationsMap.getFunctionsWithAnnotation<ComponentTimeoutHandler>()
+        functionAnnotationsMap.getFunctionsWithAnnotation<GroupTimeoutHandler>()
             .requiredFilter(FunctionFilter.nonStatic())
-            .requiredFilter(FunctionFilter.firstArg(ComponentTimeoutData::class))
+            .requiredFilter(FunctionFilter.firstArg(GroupTimeoutData::class))
             .associate {
-                it.function.findAnnotation<ComponentTimeoutHandler>()!!.name to it.toMemberFunction()
+                it.function.findAnnotation<GroupTimeoutHandler>()!!.name to it.toMemberFunction()
             }
 
     override operator fun get(handlerName: String) = map[handlerName]
